+++ Package/CHANGELOG.md
@@ -1,3 +1,12 @@
+## Version 0.1.2
+
+Features
+- Added credits worth value for shop items
+- Added normalized time to leave for enemies
+
+Fixes
+- Fixed an error that would occur if a moon had certain characters in it's name
+
 ## Version 0.1.1
 
 - Removed accidental dependency on LobbyCompatability (oops)